fix: resolve PT radio button and KC textarea test failures
- casedata-formfield-renderer: use controlled value via watch()
  instead of defaultValue for RadioButton.Group — defaultValue
  only evaluates at mount before useEffect sets form values
- support-errand-basics-about-form: use Controller for
  contactReasonDescription textarea — register with ref loses
  sync on form reset causing typed text to disappear
